+++ src/main/java/com/clothes_shop/clothes_shop/service/ProductService.java
@@ -6,10 +6,10 @@ import com.clothes_shop.clothes_shop.dto.ProductDto;
 import java.util.List;
 
 public interface ProductService {
-    Product getProductById(int id);
+    Product getProductById(Long id);
     List<Product> getAllProducts();
-    String deleteProduct(int id);
-    Product updateProduct(int id, ProductDto dto);
+    String deleteProduct(Long id);
+    Product updateProduct(Long id, ProductDto dto);
     Product addProduct(ProductDto dto);
 
 }
